+++ vector/src/main/java/im/vector/app/AppStateHandler.kt
@@ -21,8 +21,13 @@ import androidx.lifecycle.LifecycleOwner
 import arrow.core.Option
 import im.vector.app.core.di.ActiveSessionHolder
 import im.vector.app.core.utils.BehaviorDataSource
+import im.vector.app.features.home.room.detail.timeline.helper.MatrixItemColorProvider
 import im.vector.app.features.session.coroutineScope
 import im.vector.app.features.ui.UiStateRepository
+import io.reactivex.Observable
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.disposables.CompositeDisposable
+import io.reactivex.rxkotlin.addTo
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
@@ -33,8 +38,11 @@ import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 import org.matrix.android.sdk.api.extensions.tryOrNull
 import org.matrix.android.sdk.api.session.Session
+import org.matrix.android.sdk.api.session.accountdata.UserAccountDataTypes
+import org.matrix.android.sdk.api.session.events.model.toModel
 import org.matrix.android.sdk.api.session.group.model.GroupSummary
 import org.matrix.android.sdk.api.session.room.model.RoomSummary
+import org.matrix.android.sdk.rx.rx
 import javax.inject.Inject
 import javax.inject.Singleton
 
@@ -54,6 +62,7 @@ fun RoomGroupingMethod.group() = (this as? RoomGroupingMethod.ByLegacyGroup)?.gr
 @Singleton
 class AppStateHandler @Inject constructor(
         private val sessionDataSource: ActiveSessionDataSource,
+        private val matrixItemColorProvider: MatrixItemColorProvider,
         private val uiStateRepository: UiStateRepository,
         private val activeSessionHolder: ActiveSessionHolder
 ) : DefaultLifecycleObserver {
@@ -132,6 +141,11 @@ class AppStateHandler @Inject constructor(
         return (selectedSpaceDataSource.currentValue?.orNull() as? RoomGroupingMethod.ByLegacyGroup)?.groupSummary?.groupId
     }
 
+    @OnLifecycleEvent(Lifecycle.Event.ON_RESUME)
+    fun entersForeground() {
+        observeUserAccountData()
+    }
+
     override fun onResume(owner: LifecycleOwner) {
         observeActiveSession()
     }
@@ -149,5 +163,20 @@ class AppStateHandler @Inject constructor(
                 uiStateRepository.storeSelectedGroup(currentMethod.groupSummary?.groupId, session.sessionId)
             }
         }
+    }
+
+    private fun observeUserAccountData() {
+        sessionDataSource.observe()
+                .observeOn(AndroidSchedulers.mainThread())
+                .switchMap {
+                    it.orNull()?.rx()?.liveUserAccountData(setOf(UserAccountDataTypes.TYPE_OVERRIDE_COLORS))
+                            ?: Observable.just(emptyList())
+                }
+                .distinctUntilChanged()
+                .subscribe {
+                    val overrideColorSpecs = it?.firstOrNull()?.content?.toModel<Map<String, String>>()
+                    matrixItemColorProvider.setOverrideColors(overrideColorSpecs)
+                }
+                .addTo(compositeDisposable)
     }
 }